refactor: extract response-side Ctx methods into ctx_response.go
