+++ context.go
@@ -1,37 +1,16 @@
 package kruda
 
 import (
-	"bytes"
 	"context"
-	"errors"
-	"fmt"
-	"io"
 	"log/slog"
 	"mime/multipart"
 	"net/http"
-	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
 	"github.com/go-kruda/kruda/transport"
 )
-
-// jsonBufPool pools bytes.Buffer for JSON marshal output.
-// The streaming encoder writes directly into this buffer, avoiding the
-// fresh []byte allocation that Marshal() creates each call.
-// Typical JSON API response is 50–500 bytes; the buffer grows as needed
-// and is reused across requests.
-var jsonBufPool = sync.Pool{
-	New: func() any {
-		return bytes.NewBuffer(make([]byte, 0, 512))
-	},
-}
-
-// jsonContentType is a pre-allocated byte slice for the JSON content type header.
-// Used by both the fasthttp fast path (SetContentTypeBytes) and generic paths.
-var jsonContentType = []byte("application/json; charset=utf-8")
 
 // headerIntern caches canonical header keys to reduce allocations.
 // Capped at maxHeaderInternEntries to prevent memory DoS from adversarial
@@ -315,509 +294,6 @@ func (c *Ctx) cleanup() {
 	c.logger = nil
 }
 
-// Status sets the HTTP response status code. Chainable.
-func (c *Ctx) Status(code int) *Ctx {
-	c.status = code
-	return c
-}
-
-// StatusCode returns the current response status code.
-func (c *Ctx) StatusCode() int {
-	return c.status
-}
-
-// Responded returns whether a response has already been written.
-func (c *Ctx) Responded() bool {
-	return c.responded
-}
-
-// JSON sends a JSON response using the configured JSON encoder.
-// On the fasthttp path, uses SetBodyRaw for zero-copy response writing:
-// sonic.Marshal allocates a []byte, SetBodyRaw references it (no memcpy),
-// and fasthttp writes it to the TCP socket before GC reclaims it.
-// This eliminates jsonBufPool contention under high concurrency.
-// On net/http or non-default encoder paths, falls back to pooled buffer or direct marshal.
-func (c *Ctx) JSON(v any) error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-
-	// Ultra-fast path: fasthttp + default encoder → zero-copy via SetBodyRaw.
-	if c.tryFastHTTPJSONDirect(v) {
-		return nil
-	}
-
-	// Wing fast path: Marshal → SetJSON directly, skip pooled buffer.
-	if jr, ok := c.writer.(transport.JSONResponder); ok {
-		if len(c.cookies) == 0 && len(c.respHeaders) == 0 {
-			data, err := c.app.config.JSONEncoder(v)
-			if err != nil {
-				return err
-			}
-			c.responded = true
-			jr.SetJSON(c.status, data)
-			return nil
-		}
-	}
-
-	// Fast path: stream into pooled buffer (net/http transport)
-	if enc := c.app.config.JSONStreamEncoder; enc != nil {
-		return c.jsonPooled(v, enc)
-	}
-
-	// Fallback: custom JSONEncoder (no pooling)
-	data, err := c.app.config.JSONEncoder(v)
-	if err != nil {
-		return err
-	}
-	return c.jsonSend(data)
-}
-
-// jsonPooled marshals v into a pooled bytes.Buffer and writes the response.
-// The buffer is returned to the pool after the write completes.
-// The fasthttp path is inlined here to avoid the jsonSend function call overhead.
-func (c *Ctx) jsonPooled(v any, enc func(buf *bytes.Buffer, v any) error) error {
-	buf := jsonBufPool.Get().(*bytes.Buffer)
-	buf.Reset()
-
-	if err := enc(buf, v); err != nil {
-		jsonBufPool.Put(buf)
-		return err
-	}
-
-	data := buf.Bytes()
-
-	// Fast path: tryFastHTTPJSON writes directly to fasthttp, bypassing transport interface.
-	if c.tryFastHTTPJSON(data) {
-		jsonBufPool.Put(buf)
-		return nil
-	}
-
-	// Non-fasthttp path — delegate to shared jsonSend
-	err := c.jsonSend(data)
-	jsonBufPool.Put(buf)
-	return err
-}
-
-// jsonSend writes a JSON response body. data must remain valid until this returns.
-func (c *Ctx) jsonSend(data []byte) error {
-	// Fast path for fasthttp — bypass transport interface and pool-copy overhead
-	if c.tryFastHTTPJSON(data) {
-		return nil
-	}
-	// Fast path for Wing transport — bypass header interface entirely
-	if jr, ok := c.writer.(transport.JSONResponder); ok {
-		c.responded = true
-		jr.SetJSON(c.status, data)
-		return nil
-	}
-	// Fast path for net/http embedded adapter — bypass transport interface
-	if w := &c.embeddedResp; w.w != nil {
-		c.responded = true
-		h := w.w.Header()
-		if w.contentTypeSlice == nil {
-			w.contentTypeSlice = make([]string, 1)
-		}
-		w.contentTypeSlice[0] = "application/json; charset=utf-8"
-		h["Content-Type"] = w.contentTypeSlice
-		cl := len(data)
-		if cl < len(contentLengthStrings) {
-			h["Content-Length"] = []string{contentLengthStrings[cl]}
-		} else {
-			h["Content-Length"] = []string{strconv.Itoa(cl)}
-		}
-		w.w.WriteHeader(c.status)
-		w.written = true
-		_, _ = w.w.Write(data)
-		return nil
-	}
-	c.contentType = "application/json; charset=utf-8"
-	return c.sendBytes(data)
-}
-
-// Text sends a plain text response.
-func (c *Ctx) Text(s string) error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-	// Fast path for fasthttp - bypass transport interface
-	if c.tryFastHTTPText(s) {
-		return nil
-	}
-	// Fast path for transports with pre-built static response support (e.g., Wing)
-	if sr, ok := c.writer.(transport.StaticTextResponder); ok {
-		c.responded = true
-		sr.SetStaticText(c.status, "text/plain; charset=utf-8", s)
-		return nil
-	}
-	// Fast path for net/http embedded adapter - bypass transport interface
-	if w := &c.embeddedResp; w.w != nil {
-		c.responded = true
-		h := w.w.Header()
-		if w.contentTypeSlice == nil {
-			w.contentTypeSlice = make([]string, 1)
-		}
-		w.contentTypeSlice[0] = "text/plain; charset=utf-8"
-		h["Content-Type"] = w.contentTypeSlice
-		w.w.WriteHeader(c.status)
-		w.written = true
-		_, _ = io.WriteString(w.w, s)
-		return nil
-	}
-	c.contentType = "text/plain; charset=utf-8"
-	return c.sendBytes([]byte(s))
-}
-
-// HTML sends an HTML response (raw string, no template).
-// For template rendering, use c.Render() with a ViewEngine.
-func (c *Ctx) HTML(html string) error {
-	c.contentType = "text/html; charset=utf-8"
-	return c.sendBytes([]byte(html))
-}
-
-// Render renders a named template with data using the configured ViewEngine.
-//
-//	app := kruda.New(kruda.WithViews(kruda.NewViewEngine("views/*.html")))
-//	app.Get("/", func(c *kruda.Ctx) error {
-//	    return c.Render("index", Map{"title": "Home"})
-//	})
-func (c *Ctx) Render(name string, data any, code ...int) error {
-	if c.app.config.Views == nil {
-		return NewError(500, "no view engine configured")
-	}
-	if len(code) > 0 {
-		c.status = code[0]
-	}
-	buf := viewBufPool.Get().(*bytes.Buffer)
-	buf.Reset()
-	defer viewBufPool.Put(buf)
-	if err := c.app.config.Views.Render(buf, name, data); err != nil {
-		return err
-	}
-	c.contentType = "text/html; charset=utf-8"
-	return c.sendBytes(buf.Bytes())
-}
-
-// httpResponseWriter is an interface for transport.ResponseWriter implementations
-// that wrap an underlying http.ResponseWriter (e.g. netHTTPResponseWriter).
-type httpResponseWriter interface {
-	Unwrap() http.ResponseWriter
-}
-
-// File sends a file as the response.
-func (c *Ctx) File(path string) error {
-	raw := c.request.RawRequest()
-	if r, ok := raw.(*http.Request); ok {
-		// Try to unwrap the underlying http.ResponseWriter from the transport adapter.
-		if unwrapper, ok := c.writer.(httpResponseWriter); ok {
-			c.responded = true
-			http.ServeFile(unwrapper.Unwrap(), r, path)
-			return nil
-		}
-	}
-	return InternalError("file serving requires net/http transport, use NetHTTP() option")
-}
-
-// Stream sends a streaming response from an io.Reader.
-func (c *Ctx) Stream(reader io.Reader) error {
-	c.writeHeaders()
-	c.writer.WriteHeader(c.status)
-	c.responded = true
-	_, err := io.Copy(writerAdapter{c.writer}, reader)
-	return err
-}
-
-// NoContent sends a 204 No Content response.
-func (c *Ctx) NoContent() error {
-	c.status = 204
-	return c.send()
-}
-
-// SetBody sets the response body from a []byte buffer without copying (lazy-send).
-//
-// The framework flushes the body during post-handler processing — no data is
-// written to the transport until the handler returns. This means the caller
-// MUST keep the buffer valid until the handler returns; do NOT return the
-// buffer to a sync.Pool before the handler completes.
-//
-// If the handler also calls SendBytes, JSON, or Text, the eager-send wins
-// (responded flag is set) and the SetBody data is silently discarded during
-// the post-handler flush. Calling SetBody multiple times overwrites the
-// previous value (last write wins).
-//
-// Returns *Ctx for method chaining:
-//
-//	c.SetContentType("application/json").SetBody(buf)
-func (c *Ctx) SetBody(data []byte) *Ctx {
-	c.dirty |= dirtyBody
-	c.body = data
-	return c
-}
-
-// SetContentType sets the Content-Type response header.
-//
-// Returns *Ctx for method chaining:
-//
-//	c.SetContentType("text/html; charset=utf-8").SetBody(html)
-func (c *Ctx) SetContentType(ct string) *Ctx {
-	c.contentType = ct
-	return c
-}
-
-// SendBytesWithType is an optimized variant of SendBytes that sets Content-Type
-// and writes the body in a single operation. On the fasthttp path this bypasses
-// writeHeadersFastHTTP entirely — no branch checks for cacheControl, location,
-// respHeaders, or cookies. Use this for handlers that only need Content-Type
-// (e.g. json, plaintext, cached-queries).
-//
-// Precondition: caller must have already set the Date header via SetHeaderBytes.
-func (c *Ctx) SendBytesWithType(contentType string, data []byte) error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-	c.responded = true
-	c.contentLength = len(data)
-
-	// Ultra-fast path: fasthttp — bypass writeHeadersFastHTTP entirely
-	if c.trySendBytesWithTypeFastHTTP(contentType, data) {
-		return nil
-	}
-
-	// net/http and generic transport paths
-	c.contentType = contentType
-	c.writeHeaders()
-	c.writer.WriteHeader(c.status)
-	_, err := c.writer.Write(data)
-	return err
-}
-
-// SendBytesWithTypeBytes is the zero-alloc variant of SendBytesWithType.
-// Takes content-type as []byte to avoid string→[]byte conversion on the fasthttp hot path.
-// Use pre-computed []byte content-type constants for maximum performance.
-func (c *Ctx) SendBytesWithTypeBytes(contentType []byte, data []byte) error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-	c.responded = true
-
-	// Ultra-fast path: fasthttp — zero-alloc content-type write
-	if c.trySendBytesWithTypeBytesFastHTTP(contentType, data) {
-		return nil
-	}
-
-	// net/http and generic transport paths — convert back to string
-	c.contentLength = len(data)
-	c.contentType = string(contentType)
-	c.writeHeaders()
-	c.writer.WriteHeader(c.status)
-	_, err := c.writer.Write(data)
-	return err
-}
-
-// SendStaticWithTypeBytes is the zero-copy variant of SendBytesWithTypeBytes for
-// static, pre-allocated response bodies. On the fasthttp path it uses SetBodyRaw
-// instead of SetBody, avoiding a memcopy entirely.
-//
-// SAFETY: The data slice MUST be immutable for the lifetime of the program
-// (e.g. a package-level var or const-like []byte). Do NOT use with pooled
-// buffers or any data that may be modified after this call returns.
-func (c *Ctx) SendStaticWithTypeBytes(contentType []byte, data []byte) error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-	c.responded = true
-
-	// Ultra-fast path: fasthttp — zero-copy body (no memcopy)
-	if c.trySendStaticWithTypeBytesFastHTTP(contentType, data) {
-		return nil
-	}
-
-	// net/http and generic transport paths — same as SendBytesWithTypeBytes
-	c.contentLength = len(data)
-	c.contentType = string(contentType)
-	c.writeHeaders()
-	c.writer.WriteHeader(c.status)
-	_, err := c.writer.Write(data)
-	return err
-}
-
-// SendBytes writes the response body and triggers an immediate flush to the
-// transport (eager-send). Unlike SetBody, the data is written before SendBytes
-// returns, so the caller may safely return the buffer to a sync.Pool
-// immediately after the call.
-//
-// SendBytes is terminal: it returns an error and does not support method
-// chaining. If the response has already been sent (via a prior SendBytes,
-// JSON, Text, etc.), it returns ErrAlreadyResponded.
-//
-// On the fasthttp serve path the data is written via the embedded fasthttp
-// adapter for zero-copy performance. On net/http and generic transport paths
-// the data is copied by the underlying http.ResponseWriter.
-func (c *Ctx) SendBytes(data []byte) error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-	c.responded = true
-	c.contentLength = len(data)
-
-	// Fast path: fasthttp — use embedded adapter directly (c.writer is nil)
-	if c.trySendBytesFastHTTP(data) {
-		return nil
-	}
-
-	// net/http and generic transport paths — use existing writeHeaders + writer
-	c.writeHeaders()
-	c.writer.WriteHeader(c.status)
-	_, err := c.writer.Write(data)
-	return err
-}
-
-// Redirect sends a redirect response. Default status is 302.
-func (c *Ctx) Redirect(url string, code ...int) error {
-	c.body = nil // Clear any SetBody data — redirects MUST NOT have a body
-	status := 302
-	if len(code) > 0 {
-		status = code[0]
-	}
-	c.status = status
-	c.location = url // fixed slot — no map write
-	return c.send()
-}
-
-// sanitizeHeaderValue strips CR and LF characters from a header value to prevent
-// HTTP header injection (CRLF injection). Most values pass through unchanged via
-// the fast path check.
-func sanitizeHeaderValue(value string) string {
-	if !strings.ContainsAny(value, "\r\n") {
-		return value
-	}
-	var b strings.Builder
-	b.Grow(len(value))
-	for i := 0; i < len(value); i++ {
-		if value[i] != '\r' && value[i] != '\n' {
-			b.WriteByte(value[i])
-		}
-	}
-	return b.String()
-}
-
-// isValidHeaderKey checks if key contains only valid token characters per RFC 7230.
-// token = 1*tchar
-// tchar = "!" / "#" / "$" / "%" / "&" / "'" / "*" / "+" / "-" / "." /
-//
-//	"^" / "_" / "`" / "|" / "~" / DIGIT / ALPHA
-func isValidHeaderKey(key string) bool {
-	if len(key) == 0 {
-		return false
-	}
-	for i := 0; i < len(key); i++ {
-		if !isTokenChar(key[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-// isTokenChar returns true if c is a valid HTTP token character per RFC 7230.
-func isTokenChar(c byte) bool {
-	if (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z') || (c >= '0' && c <= '9') {
-		return true
-	}
-	switch c {
-	case '!', '#', '$', '%', '&', '\'', '*', '+', '-', '.', '^', '_', '`', '|', '~':
-		return true
-	}
-	return false
-}
-
-// SetHeader sets a response header, replacing any existing values. Chainable.
-// Common headers (Content-Type, Cache-Control, Location) use fixed slots
-// to avoid map allocation on the hot path.
-// Validates key per RFC 7230 and strips CRLF from value to prevent header injection.
-func (c *Ctx) SetHeader(key, value string) *Ctx {
-	// Validate key on all paths — fasthttp has its own removeNewLines for values,
-	// but does not reject invalid keys (it normalizes them instead).
-	if !isValidHeaderKey(key) {
-		c.app.config.Logger.Warn("kruda: invalid header key, skipping", "key", key)
-		return c
-	}
-	// Fast path: write directly to fasthttp response headers.
-	// fasthttp's Set() calls removeNewLines() internally, so CRLF injection
-	// is prevented by fasthttp itself (replaces \r\n with space).
-	if c.trySetHeaderFastHTTP(key, value) {
-		return c
-	}
-	value = sanitizeHeaderValue(value)
-	switch key {
-	case "Content-Type":
-		c.contentType = value
-	case "Cache-Control":
-		c.cacheControl = value
-	case "Location":
-		c.location = value
-	default:
-		c.dirty |= dirtyRespHdrs
-		c.respHeaders[internHeader(key)] = []string{value}
-	}
-	return c
-}
-
-// SetHeaderBytes sets a response header with a []byte value.
-// On the fasthttp path this is zero-alloc (avoids []byte→string conversion).
-// Falls back to SetHeader(key, string(value)) on net/http.
-func (c *Ctx) SetHeaderBytes(key string, value []byte) *Ctx {
-	if !isValidHeaderKey(key) {
-		c.app.config.Logger.Warn("kruda: invalid header key, skipping", "key", key)
-		return c
-	}
-	if c.trySetHeaderBytesFastHTTP(key, value) {
-		return c
-	}
-	return c.SetHeader(key, string(value))
-}
-
-// AddHeader appends a value to a response header without replacing existing values.
-// Supports multi-value headers like Vary, Cache-Control. Chainable.
-// Fixed-slot headers: Content-Type and Location always replace (single-valued per RFC).
-// Cache-Control appends comma-separated (multi-valued per RFC 7234 section 5.2).
-// Validates key per RFC 7230 and strips CRLF from value to prevent header injection.
-func (c *Ctx) AddHeader(key, value string) *Ctx {
-	if !isValidHeaderKey(key) {
-		c.app.config.Logger.Warn("kruda: invalid header key, skipping", "key", key)
-		return c
-	}
-	value = sanitizeHeaderValue(value)
-	switch key {
-	case "Content-Type":
-		c.contentType = value
-	case "Cache-Control":
-		if c.cacheControl != "" {
-			c.cacheControl += ", " + value
-		} else {
-			c.cacheControl = value
-		}
-	case "Location":
-		c.location = value
-	default:
-		key = internHeader(key)
-		c.dirty |= dirtyRespHdrs
-		c.respHeaders[key] = append(c.respHeaders[key], value)
-	}
-	return c
-}
-
-// SetCookie sets a cookie on the response. Supports multiple cookies.
-// Cookie values are sanitized to prevent header injection.
-// Supports SameSite attribute and MaxAge<=0 for deletion.
-func (c *Ctx) SetCookie(cookie *Cookie) *Ctx {
-	cookie.Value = sanitizeHeaderValue(cookie.Value)
-	cookie.Path = sanitizeHeaderValue(cookie.Path)
-	cookie.Domain = sanitizeHeaderValue(cookie.Domain)
-	c.dirty |= dirtyCookies
-	c.cookies = append(c.cookies, cookie)
-	return c
-}
-
 // Set stores a value in the request-scoped locals.
 func (c *Ctx) Set(key string, value any) {
 	c.dirty |= dirtyLocals
@@ -942,312 +418,6 @@ func Need[T any](c *Ctx, key string) (T, bool) {
 		return zero, false
 	}
 	return typed, true
-}
-
-// ErrAlreadyResponded is returned when a handler attempts to write a response
-// after one has already been sent. Check with errors.Is(err, ErrAlreadyResponded).
-var ErrAlreadyResponded = fmt.Errorf("kruda: response already sent")
-
-func (c *Ctx) send() error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-	c.responded = true
-
-	// Per HTTP spec, 204 No Content MUST NOT have a body.
-	if c.status == 204 {
-		c.body = nil
-	}
-	if c.body != nil {
-		c.contentLength = len(c.body)
-	}
-
-	// Fast path: fasthttp — use embedded adapter directly (c.writer is nil)
-	if c.trySendFastHTTP() {
-		return nil
-	}
-
-	// net/http and generic transport paths
-	c.writeHeaders()
-	c.writer.WriteHeader(c.status)
-	if c.body != nil {
-		_, err := c.writer.Write(c.body)
-		c.body = nil
-		return err
-	}
-	return nil
-}
-
-func (c *Ctx) sendBytes(data []byte) error {
-	if c.responded {
-		return ErrAlreadyResponded
-	}
-	c.responded = true
-	c.contentLength = len(data)
-
-	// Fast path: fasthttp — use embedded adapter directly (c.writer is nil)
-	if c.trySendBytesFastHTTP(data) {
-		return nil
-	}
-
-	c.writeHeaders()
-	c.writer.WriteHeader(c.status)
-
-	// Direct write — transport.ResponseWriter implementations copy internally.
-	// On the net/http path, http.ResponseWriter.Write already copies to its
-	// internal buffer, so a pooled intermediate buffer would be a redundant copy.
-	_, err := c.writer.Write(data)
-	return err
-}
-
-// contentLengthStrings is a pre-computed lookup table for common Content-Length
-// values as strings, avoiding strconv.Itoa allocation on the net/http path.
-// Covers 0–2048 which handles the majority of JSON API responses.
-var contentLengthStrings [2049]string
-
-func init() {
-	for i := range contentLengthStrings {
-		contentLengthStrings[i] = strconv.Itoa(i)
-	}
-}
-
-func contentLengthString(length int) string {
-	if length >= 0 && length < len(contentLengthStrings) {
-		return contentLengthStrings[length]
-	}
-	return strconv.Itoa(length)
-}
-
-func (c *Ctx) writeHeaders() {
-	// Fast path: net/http embedded adapter — boolean check instead of type assert
-	if c.embeddedResp.w != nil {
-		// Simple case: only content-type + content-length, no other headers
-		if c.cacheControl == "" && c.location == "" &&
-			len(c.respHeaders) == 0 && len(c.cookies) == 0 &&
-			len(c.app.secHeaders) == 0 {
-			if c.contentType != "" {
-				c.embeddedResp.SetContentType(c.contentType)
-			}
-			if c.contentLength >= 0 {
-				c.embeddedResp.SetContentLength(contentLengthString(c.contentLength))
-			}
-			return
-		}
-		// Complex case: use direct header access
-		c.writeHeadersNetHTTP()
-		return
-	}
-
-	// Fallback: custom transport — use interface methods
-	c.writeHeadersGeneric()
-}
-
-// writeHeadersNetHTTP writes all response headers using the embedded net/http
-// adapter's direct header access. Called when the response has custom headers,
-// cookies, or security headers beyond simple content-type/content-length.
-func (c *Ctx) writeHeadersNetHTTP() {
-	// Use embedded adapter's optimized methods for common headers
-	if c.contentType != "" {
-		c.embeddedResp.SetContentType(c.contentType)
-	}
-	if c.contentLength >= 0 {
-		c.embeddedResp.SetContentLength(contentLengthString(c.contentLength))
-	}
-
-	// Direct header access for other headers
-	h := c.embeddedResp.DirectHeader()
-
-	if c.cacheControl != "" {
-		h.Set("Cache-Control", c.cacheControl)
-	}
-	if c.location != "" {
-		h.Set("Location", c.location)
-	}
-
-	if len(c.respHeaders) > 0 {
-		for k, vals := range c.respHeaders {
-			h[k] = vals
-		}
-	}
-
-	if len(c.cookies) > 0 {
-		for _, cookie := range c.cookies {
-			h.Add("Set-Cookie", formatCookie(cookie))
-		}
-	}
-
-	if len(c.app.secHeaders) > 0 {
-		for _, kv := range c.app.secHeaders {
-			if h[kv[0]] == nil {
-				h[kv[0]] = []string{kv[1]}
-			}
-		}
-	}
-}
-
-// writeHeadersGeneric writes all response headers using the transport interface
-// methods. Used as fallback for custom transports that are not the embedded
-// net/http adapter.
-func (c *Ctx) writeHeadersGeneric() {
-	h := c.writer.Header()
-
-	if c.contentType != "" {
-		h.Set("Content-Type", c.contentType)
-	}
-	if c.contentLength >= 0 {
-		h.Set("Content-Length", contentLengthString(c.contentLength))
-	}
-	if c.cacheControl != "" {
-		h.Set("Cache-Control", c.cacheControl)
-	}
-	if c.location != "" {
-		h.Set("Location", c.location)
-	}
-
-	if len(c.respHeaders) > 0 {
-		for k, vals := range c.respHeaders {
-			if len(vals) == 1 {
-				h.Set(k, vals[0])
-			} else {
-				for i, v := range vals {
-					if i == 0 {
-						h.Set(k, v)
-					} else {
-						h.Add(k, v)
-					}
-				}
-			}
-		}
-	}
-
-	if len(c.cookies) > 0 {
-		for _, cookie := range c.cookies {
-			h.Add("Set-Cookie", formatCookie(cookie))
-		}
-	}
-
-	if len(c.app.secHeaders) > 0 {
-		if directAccess, ok := h.(transport.DirectHeaderAccess); ok {
-			if httpHeader := directAccess.DirectHeader(); httpHeader != nil {
-				for _, kv := range c.app.secHeaders {
-					if httpHeader[kv[0]] == nil {
-						httpHeader[kv[0]] = []string{kv[1]}
-					}
-				}
-			} else {
-				for _, kv := range c.app.secHeaders {
-					if h.Get(kv[0]) == "" {
-						h.Set(kv[0], kv[1])
-					}
-				}
-			}
-		} else {
-			for _, kv := range c.app.secHeaders {
-				if h.Get(kv[0]) == "" {
-					h.Set(kv[0], kv[1])
-				}
-			}
-		}
-	}
-}
-
-// formatCookie serializes a Cookie to a Set-Cookie header value.
-// Sanitizes name and value to prevent header injection. Supports SameSite and MaxAge<0 for deletion.
-func formatCookie(cookie *Cookie) string {
-	name := sanitizeCookieToken(cookie.Name)
-	value := sanitizeCookieValue(cookie.Value)
-
-	var b strings.Builder
-	b.WriteString(name)
-	b.WriteByte('=')
-	b.WriteString(value)
-
-	if cookie.Path != "" {
-		b.WriteString("; Path=")
-		b.WriteString(sanitizeCookieValue(cookie.Path))
-	}
-	if cookie.Domain != "" {
-		b.WriteString("; Domain=")
-		b.WriteString(sanitizeCookieValue(cookie.Domain))
-	}
-	if !cookie.Expires.IsZero() {
-		b.WriteString("; Expires=")
-		b.WriteString(cookie.Expires.UTC().Format(http.TimeFormat))
-	}
-	// MaxAge < 0 means delete cookie (set Max-Age=0)
-	if cookie.MaxAge > 0 {
-		b.WriteString("; Max-Age=")
-		b.WriteString(strconv.Itoa(cookie.MaxAge))
-	} else if cookie.MaxAge < 0 {
-		b.WriteString("; Max-Age=0")
-	}
-	if cookie.Secure {
-		b.WriteString("; Secure")
-	}
-	if cookie.HTTPOnly {
-		b.WriteString("; HttpOnly")
-	}
-	switch cookie.SameSite {
-	case http.SameSiteLaxMode:
-		b.WriteString("; SameSite=Lax")
-	case http.SameSiteStrictMode:
-		b.WriteString("; SameSite=Strict")
-	case http.SameSiteNoneMode:
-		b.WriteString("; SameSite=None")
-	}
-	return b.String()
-}
-
-// sanitizeCookieToken removes characters not allowed in cookie names (RFC 6265 token).
-func sanitizeCookieToken(s string) string {
-	var b strings.Builder
-	for i := 0; i < len(s); i++ {
-		c := s[i]
-		if c > 0x20 && c < 0x7f && !isCookieSeparator(c) {
-			b.WriteByte(c)
-		}
-	}
-	return b.String()
-}
-
-// sanitizeCookieValue removes characters not allowed in cookie values.
-// Strips semicolons, newlines, and other control characters to prevent header injection.
-func sanitizeCookieValue(s string) string {
-	var b strings.Builder
-	for i := 0; i < len(s); i++ {
-		c := s[i]
-		if c == 0x21 || (c >= 0x23 && c <= 0x2b) || (c >= 0x2d && c <= 0x3a) || (c >= 0x3c && c <= 0x5b) || (c >= 0x5d && c <= 0x7e) {
-			b.WriteByte(c)
-		}
-	}
-	return b.String()
-}
-
-func isCookieSeparator(c byte) bool {
-	switch c {
-	case '(', ')', '<', '>', '@', ',', ';', ':', '\\', '"', '/', '[', ']', '?', '=', '{', '}', ' ', '\t':
-		return true
-	}
-	return false
-}
-
-// writerAdapter adapts transport.ResponseWriter to io.Writer for io.Copy.
-type writerAdapter struct {
-	w transport.ResponseWriter
-}
-
-func (a writerAdapter) Write(p []byte) (int, error) {
-	return a.w.Write(p)
-}
-
-// isBodyTooLarge checks if an error indicates the request body exceeded the size limit.
-// Works with transport.ErrBodyTooLarge from both net/http and fasthttp transports.
-func isBodyTooLarge(err error) bool {
-	if err == nil {
-		return false
-	}
-	var btle *transport.BodyTooLargeError
-	return errors.As(err, &btle)
 }
 
 // Map is a convenience type alias for map[string]any.
